+++ send-surveys/src/main/java/com/objectcomputing/pulsesurvey/repositories/ResponseKeyRepository.java
@@ -1,4 +1,4 @@
-package repositories;
+package com.objectcomputing.pulsesurvey.repositories;
 
 import java.util.List;
 import java.util.UUID;
@@ -6,7 +6,7 @@ import java.util.UUID;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
-import model.ResponseKey;
+import com.objectcomputing.pulsesurvey.model.ResponseKey;
 import io.micronaut.data.jdbc.annotation.JdbcRepository;
 import io.micronaut.data.model.query.builder.sql.Dialect;
 import io.micronaut.data.repository.CrudRepository;
