+++ send-surveys/src/main/java/com/objectcomputing/pulsesurvey/template/manager/MustacheConfig.java
@@ -1,4 +1,4 @@
-package template.mgr;
+package com.objectcomputing.pulsesurvey.template.manager;
 
 import com.github.mustachejava.DefaultMustacheFactory;
 import com.github.mustachejava.MustacheFactory;
