+++ send-surveys/src/main/java/com/objectcomputing/pulsesurvey/template/manager/SurveyTemplateManager.java
@@ -1,4 +1,4 @@
-package template.mgr;
+package com.objectcomputing.pulsesurvey.template.manager;
 
 import com.github.mustachejava.Mustache;
 import com.github.mustachejava.MustacheFactory;
